@@ -20,46 +20,46 @@
 	},
 	"dependencies": {
 		"@floating-ui/dom": "1.6.12",
-		"@tanstack/svelte-query": "^5.62.3",
+		"@tanstack/svelte-query": "^5.62.8",
 		"bcryptjs": "^2.4.3",
 		"cuid": "^3.0.0",
-		"highlight.js": "11.10.0",
-		"mongodb": "^6.11.0",
-		"svelte-awesome": "^3.3.4",
-		"uuid": "^10.0.0"
+		"highlight.js": "11.11.0",
+		"mongodb": "^6.12.0",
+		"svelte-awesome": "^3.3.5",
+		"uuid": "^11.0.3"
 	},
 	"devDependencies": {
-		"@playwright/test": "^1.48.2",
-		"@skeletonlabs/skeleton": "2.10.3",
+		"@playwright/test": "^1.49.1",
+		"@skeletonlabs/skeleton": "2.10.4",
 		"@skeletonlabs/tw-plugin": "0.4.0",
-		"@sveltejs/adapter-auto": "^3.2.5",
-		"@sveltejs/kit": "^2.9.1",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.12.2",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@tailwindcss/forms": "0.5.9",
 		"@tailwindcss/typography": "0.5.15",
 		"@types/bcryptjs": "^2.4.6",
 		"@types/eslint": "^9.6.1",
-		"@types/node": "22.9.0",
+		"@types/node": "22.10.2",
 		"@types/uuid": "^10.0.0",
 		"autoprefixer": "10.4.20",
 		"cookie": "^1.0.2",
-		"eslint": "^9.16.0",
+		"eslint": "^9.17.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.46.1",
-		"globals": "^15.11.0",
+		"globals": "^15.14.0",
 		"husky": "^9.1.7",
 		"postcss": "8.4.49",
-		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.7",
+		"prettier": "^3.4.2",
+		"prettier-plugin-svelte": "^3.3.2",
 		"set-cookie-parser": "^2.7.1",
-		"svelte": "^4.2.19",
+		"svelte": "^5.14.4",
 		"svelte-check": "^4.1.1",
-		"tailwindcss": "3.4.15",
+		"tailwindcss": "3.4.17",
 		"tslib": "^2.8.1",
 		"typescript": "^5.7.2",
-		"typescript-eslint": "^8.17.0",
-		"vite": "^5.4.11",
-		"vite-plugin-tailwind-purgecss": "0.3.3",
-		"vitest": "^2.1.4"
+		"typescript-eslint": "^8.18.1",
+		"vite": "^6.0.3",
+		"vite-plugin-tailwind-purgecss": "0.3.5",
+		"vitest": "^2.1.8"
 	}
 }
