@@ -1,9 +1,15 @@
 <script lang="ts">
+	import { preventDefault } from 'svelte/legacy';
+
 	import { createEventDispatcher } from 'svelte';
 
-	export let links: { self: string };
-	export let meta: { total_pages: number };
-	export let searchParams: { page: string; page_size: string; schema: string };
+	interface Props {
+		links: { self: string };
+		meta: { total_pages: number };
+		searchParams: { page: string; page_size: string; schema: string };
+	}
+
+	let { links, meta, searchParams }: Props = $props();
 
 	const dispatch = createEventDispatcher();
 
@@ -57,7 +63,7 @@
 				<a
 					href={`/index-explorer?${createPageLink(currentPage - 1)}`}
 					class="ml-0 rounded-l-lg border border-gray-300 bg-white px-3 py-2 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-					on:click|preventDefault={() => handlePageChange(currentPage - 1)}
+					onclick={preventDefault(() => handlePageChange(currentPage - 1))}
 				>
 					Previous
 				</a>
@@ -78,7 +84,7 @@
 						currentPage
 							? 'bg-primary-500 text-white'
 							: 'bg-white text-gray-500 dark:bg-gray-800 dark:text-gray-400'}"
-						on:click|preventDefault={() => handlePageChange(Number(page))}
+						onclick={preventDefault(() => handlePageChange(Number(page)))}
 					>
 						{page}
 					</a>
@@ -90,7 +96,7 @@
 				<a
 					href={`/index-explorer?${createPageLink(currentPage + 1)}`}
 					class="rounded-r-lg border border-gray-300 bg-white px-3 py-2 leading-tight text-gray-500 hover:bg-gray-100 hover:text-gray-700 dark:border-gray-700 dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
-					on:click|preventDefault={() => handlePageChange(currentPage + 1)}
+					onclick={preventDefault(() => handlePageChange(currentPage + 1))}
 				>
 					Next
 				</a>
